requirements.txt
psycopg2-binary==2.9.5
django-bootstrap5==22.2
Django==4.1.7
djangorestframework==3.14.0
